+++ nc_py_api/files.py
@@ -2,9 +2,6 @@
 
 import builtins
 import os
-from dataclasses import dataclass
-from datetime import datetime
-from email.utils import parsedate_to_datetime
 from io import BytesIO
 from json import dumps, loads
 from pathlib import Path
@@ -19,163 +16,8 @@ from httpx import Response
 
 from ._session import NcSessionBasic
 from .exceptions import NextcloudException, check_error
-
-
-@dataclass
-class FsNodeInfo:
-    """Extra FS object attributes from Nextcloud"""
-
-    size: int
-    """Length of file in bytes, zero for directories.."""
-    content_length: int
-    """For directories it is size of all content in it, for files it is equal to ``size``."""
-    permissions: str
-    """Permissions for the object."""
-    favorite: bool
-    """Flag indicating if the object is marked as favorite."""
-    fileid: int
-    """Clear file ID without Nextcloud instance ID."""
-    _last_modified: datetime
-
-    def __init__(self, **kwargs):
-        self.size = kwargs.get("size", 0)
-        self.content_length = kwargs.get("content_length", 0)
-        self.permissions = kwargs.get("permissions", "")
-        self.favorite = kwargs.get("favorite", False)
-        self.fileid = kwargs.get("fileid", 0)
-        try:
-            self.last_modified = kwargs.get("last_modified", datetime(1970, 1, 1))
-        except (ValueError, TypeError):
-            self.last_modified = datetime(1970, 1, 1)
-
-    @property
-    def last_modified(self) -> datetime:
-        """Time when the object was last modified.
-
-        .. note:: ETag if more preferable way to check if the object was changed."""
-
-        return self._last_modified
-
-    @last_modified.setter
-    def last_modified(self, value: Union[str, datetime]):
-        if isinstance(value, str):
-            self._last_modified = parsedate_to_datetime(value)
-        else:
-            self._last_modified = value
-
-
-@dataclass
-class FsNode:
-    """A class that represents a Nextcloud file object.
-
-    Acceptable itself as a ``path`` parameter for the most file APIs."""
-
-    full_path: str
-    """Path to the object, including the username. Does not include `dav` prefix"""
-
-    file_id: str
-    """File ID + NC instance ID"""
-
-    etag: str
-    """An entity tag (ETag) of the object"""
-
-    info: FsNodeInfo
-    """Additional extra information for the object"""
-
-    def __init__(self, full_path: str, **kwargs):
-        self.full_path = full_path
-        self.file_id = kwargs.get("file_id", "")
-        self.etag = kwargs.get("etag", "")
-        self.info = FsNodeInfo(**kwargs)
-
-    @property
-    def is_dir(self) -> bool:
-        """Returns ``True`` for the directories, ``False`` otherwise."""
-
-        return self.full_path.endswith("/")
-
-    def __str__(self):
-        return (
-            f"{'Dir' if self.is_dir else 'File'}: `{self.name}` with id={self.file_id}"
-            f" last modified at {str(self.info.last_modified)} and {self.info.permissions} permissions."
-        )
-
-    def __eq__(self, other):
-        if self.file_id and self.file_id == other.file_id:
-            return True
-        return False
-
-    @property
-    def has_extra(self) -> bool:
-        """Flag indicating whether this ``FsNode`` was obtained by the `mkdir` or `upload`
-        methods and does not contain extended information."""
-
-        return bool(self.info.permissions)
-
-    @property
-    def name(self) -> str:
-        """Returns last ``pathname`` component."""
-
-        return self.full_path.rstrip("/").rsplit("/", maxsplit=1)[-1]
-
-    @property
-    def user(self) -> str:
-        """Returns user ID extracted from the `full_path`."""
-
-        return self.full_path.lstrip("/").split("/", maxsplit=2)[1]
-
-    @property
-    def user_path(self) -> str:
-        """Returns path relative to the user's root directory."""
-
-        return self.full_path.lstrip("/").split("/", maxsplit=2)[-1]
-
-    @property
-    def is_shared(self) -> bool:
-        """Check if a file or folder is shared"""
-
-        return self.info.permissions.find("S") != -1
-
-    @property
-    def is_shareable(self) -> bool:
-        """Check if a file or folder can be shared"""
-
-        return self.info.permissions.find("R") != -1
-
-    @property
-    def is_mounted(self) -> bool:
-        """Check if a file or folder is mounted"""
-
-        return self.info.permissions.find("M") != -1
-
-    @property
-    def is_readable(self) -> bool:
-        """Check if the file or folder is readable"""
-
-        return self.info.permissions.find("G") != -1
-
-    @property
-    def is_deletable(self) -> bool:
-        """Check if a file or folder can be deleted"""
-
-        return self.info.permissions.find("D") != -1
-
-    @property
-    def is_updatable(self) -> bool:
-        """Check if file/directory is writable"""
-
-        if self.is_dir:
-            return self.info.permissions.find("NV") != -1
-        return self.info.permissions.find("W") != -1
-
-    @property
-    def is_creatable(self) -> bool:
-        """Check whether new files or folders can be created inside this folder"""
-
-        if not self.is_dir:
-            return False
-        return self.info.permissions.find("CK") != -1
-
+from .files_defs import FsNode
+from .files_sharing import FilesSharingAPI
 
 PROPFIND_PROPERTIES = [
     "d:resourcetype",
@@ -212,10 +54,14 @@ SEARCH_PROPERTIES_MAP = {
 
 
 class FilesAPI:
-    """This class provides all WebDAV functionality related to the files."""
+    """This class provides all File System functionality and File Sharing abilities."""
+
+    sharing: FilesSharingAPI
+    """API for managing Files Shares"""
 
     def __init__(self, session: NcSessionBasic):
         self._session = session
+        self.sharing = FilesSharingAPI(session)
 
     def listdir(self, path: Union[str, FsNode] = "", depth: int = 1, exclude_self=True) -> list[FsNode]:
         """Returns a list of all entries in the specified directory.
@@ -233,7 +79,7 @@ class FilesAPI:
         return self._listdir(self._session.user, path, properties=properties, depth=depth, exclude_self=exclude_self)
 
     def by_id(self, file_id: Union[int, str, FsNode]) -> Optional[FsNode]:
-        """Returns :py:class:`FsNode` by file_id if any.
+        """Returns :py:class:`~nc_py_api.files_defs.FsNode` by file_id if any.
 
         :param file_id: can be full file ID with Nextcloud instance ID or only clear file ID.
         """
@@ -243,7 +89,7 @@ class FilesAPI:
         return result[0] if result else None
 
     def by_path(self, path: Union[str, FsNode]) -> Optional[FsNode]:
-        """Returns :py:class:`FsNode` by exact path if any."""
+        """Returns :py:class:`~nc_py_api.files_defs.FsNode` by exact path if any."""
 
         path = path.user_path if isinstance(path, FsNode) else path
         result = self.listdir(path, depth=0, exclude_self=False)