+++ nc_py_api/ui_files_actions_menu.py
@@ -28,6 +28,8 @@ class UiFilesActionsAPI:
         self._session = session
 
     def register(self, name: str, display_name: str, callback_url: str, **kwargs) -> None:
+        """Registers the files a dropdown menu element."""
+
         require_capabilities("app_ecosystem_v2", self._session.capabilities)
         params = {
             "fileActionMenuParams": {
@@ -44,6 +46,8 @@ class UiFilesActionsAPI:
         self._session.ocs(method="POST", path=f"{APP_V2_BASIC_URL}/{ENDPOINT_SUFFIX}", json=params)
 
     def unregister(self, name: str, not_fail=True) -> None:
+        """Removes files dropdown menu element."""
+
         require_capabilities("app_ecosystem_v2", self._session.capabilities)
         params = {"fileActionMenuName": name}
         try: